+++ orderinformation.php
@@ -110,8 +110,6 @@
 		<div class="container">
 			<div id="shoppingcart">
                 <h1>Order Information</h1>
-                <p style = "text-align: center; font-size: 15px;">Your order has been requested. It will be processed in 5-6 working days.</p>
-
                 <?php 
                     // Prepare variabiles for db connection
                     $serverName = "localhost";
@@ -126,39 +124,45 @@
                     if($conn->connect_error) {
                         die("Connection failed".$conn->connect_error);
                     }
-                    
-                    $query = mysqli_query($conn,'SELECT * FROM cart WHERE user_id='.$_SESSION['id']);
+                    if(isset($_SESSION['id'])) {
+                        $query = mysqli_query($conn,'SELECT * FROM cart WHERE user_id='.$_SESSION['id']);
+                    }
                 ?>
 
                 <table>
                     <tr>
-                        <th>Shipping Address</th>
+                        <th>User ID</th>
                         <th>Product Name</th>
                         <th>Price</th>
                         <th>Qty</th>
                         <th>Amount</th>
                     </tr>
                     <?php
-                        $total = 0;
-                        while($row = mysqli_fetch_array($query)) { 
-                            $amount = ($row['price'] * $row['quantity']);
-                            $total += $amount;
-                            echo "<tr>
-                                    <td>".$row['user_id']."</td>
-                                    <td>".$row['product_name']."</td>
-                                    <td>"."$".$row['price']."</td>
-                                    <td>".$row['quantity']."</td>
-                                    <td>"."$".$amount."</td>
-                                </tr>";    
+                        // if(isset($_SESSION['id'])) {};
+                        // if(!isset($_SESSION['id'])) {};
+                        if(isset($_SESSION['id'])) {
+                            $total = 0;
+                            while($row = mysqli_fetch_array($query)) { 
+                                $amount = ($row['price'] * $row['quantity']);
+                                $total += $amount;
+                                echo "<tr>
+                                        <td>".$row['user_id']."</td>
+                                        <td>".$row['product_name']."</td>
+                                        <td>"."$".$row['price']."</td>
+                                        <td>".$row['quantity']."</td>
+                                        <td>"."$".$amount."</td>
+                                    </tr>";    
+                            }
+                            mysqli_close($conn);
+
+                            echo "<tr><td colspan='6'>"."Total: "."$".$total."</td></tr>";
+
+                            echo "<tr><td colspan='6'><p style = 'text-align: center; font-size: 15px;'>Your order has been requested. It will be processed in 5-6 working days until it reaches to your shipping address.</p></td></tr>";
+
+                            echo "<tr id='ordertotal'>
+                                        <td colspan='6'><button><a href='includes/deletefromcart.inc.php?user_id=".$_SESSION['id']."'>Cancel Order</a></button>
+                                    </tr>";
                         }
-                        mysqli_close($conn);
-
-                        echo "<tr><td colspan='6'>"."Total: "."$".$total."</td></tr>";
-
-                        echo "<tr id='ordertotal'>
-                                    <td colspan='6'><button><a href='includes/deletefromcart.inc.php?user_id=".$_SESSION['id']."'>Cancel Order</a></button>
-                                </tr>";
-       
                     ?>    
                 </table>
                 <button id="ContinueShopping"><a href="home.php">Continue shopping</a></button>
